reduce one traversal for shards
