+++ server/src/main/java/org/elasticsearch/cluster/routing/allocation/allocator/BalancedShardsAllocator.java
@@ -26,6 +26,8 @@ import org.apache.lucene.util.IntroSorter;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.metadata.Metadata;
 import org.elasticsearch.cluster.routing.RoutingNode;
+import org.elasticsearch.cluster.routing.RoutingNode.ModelIndex;
+import org.elasticsearch.cluster.routing.RoutingNode.ModelNode;
 import org.elasticsearch.cluster.routing.RoutingNodes;
 import org.elasticsearch.cluster.routing.ShardRouting;
 import org.elasticsearch.cluster.routing.ShardRoutingState;
@@ -53,11 +55,9 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.StreamSupport;
 
 import static org.elasticsearch.cluster.routing.ShardRoutingState.RELOCATING;
@@ -731,18 +731,7 @@ public class BalancedShardsAllocator implements ShardsAllocator {
         private Map<String, ModelNode> buildModelFromAssigned() {
             Map<String, ModelNode> nodes = new HashMap<>();
             for (RoutingNode rn : routingNodes) {
-                ModelNode node = new ModelNode(rn);
-                nodes.put(rn.nodeId(), node);
-                for (ShardRouting shard : rn) {
-                    assert rn.nodeId().equals(shard.currentNodeId());
-                    /* we skip relocating shards here since we expect an initializing shard with the same id coming in */
-                    if (shard.state() != RELOCATING) {
-                        node.addShard(shard);
-                        if (logger.isTraceEnabled()) {
-                            logger.trace("Assigned shard [{}] to node [{}]", shard, node.getNodeId());
-                        }
-                    }
-                }
+                nodes.put(rn.nodeId(), rn.modelNode());
             }
             return nodes;
         }
@@ -1012,136 +1001,6 @@ public class BalancedShardsAllocator implements ShardsAllocator {
             return false;
         }
 
-    }
-
-    static class ModelNode implements Iterable<ModelIndex> {
-        private final Map<String, ModelIndex> indices = new HashMap<>();
-        private int numShards = 0;
-        private final RoutingNode routingNode;
-
-        ModelNode(RoutingNode routingNode) {
-            this.routingNode = routingNode;
-        }
-
-        public ModelIndex getIndex(String indexId) {
-            return indices.get(indexId);
-        }
-
-        public String getNodeId() {
-            return routingNode.nodeId();
-        }
-
-        public RoutingNode getRoutingNode() {
-            return routingNode;
-        }
-
-        public int numShards() {
-            return numShards;
-        }
-
-        public int numShards(String idx) {
-            ModelIndex index = indices.get(idx);
-            return index == null ? 0 : index.numShards();
-        }
-
-        public int highestPrimary(String index) {
-            ModelIndex idx = indices.get(index);
-            if (idx != null) {
-                return idx.highestPrimary();
-            }
-            return -1;
-        }
-
-        public void addShard(ShardRouting shard) {
-            ModelIndex index = indices.get(shard.getIndexName());
-            if (index == null) {
-                index = new ModelIndex(shard.getIndexName());
-                indices.put(index.getIndexId(), index);
-            }
-            index.addShard(shard);
-            numShards++;
-        }
-
-        public void removeShard(ShardRouting shard) {
-            ModelIndex index = indices.get(shard.getIndexName());
-            if (index != null) {
-                index.removeShard(shard);
-                if (index.numShards() == 0) {
-                    indices.remove(shard.getIndexName());
-                }
-            }
-            numShards--;
-        }
-
-        @Override
-        public String toString() {
-            StringBuilder sb = new StringBuilder();
-            sb.append("Node(").append(routingNode.nodeId()).append(")");
-            return sb.toString();
-        }
-
-        @Override
-        public Iterator<ModelIndex> iterator() {
-            return indices.values().iterator();
-        }
-
-        public boolean containsShard(ShardRouting shard) {
-            ModelIndex index = getIndex(shard.getIndexName());
-            return index == null ? false : index.containsShard(shard);
-        }
-
-    }
-
-    static final class ModelIndex implements Iterable<ShardRouting> {
-        private final String id;
-        private final Set<ShardRouting> shards = new HashSet<>(4); // expect few shards of same index to be allocated on same node
-        private int highestPrimary = -1;
-
-        ModelIndex(String id) {
-            this.id = id;
-        }
-
-        public int highestPrimary() {
-            if (highestPrimary == -1) {
-                int maxId = -1;
-                for (ShardRouting shard : shards) {
-                    if (shard.primary()) {
-                        maxId = Math.max(maxId, shard.id());
-                    }
-                }
-                return highestPrimary = maxId;
-            }
-            return highestPrimary;
-        }
-
-        public String getIndexId() {
-            return id;
-        }
-
-        public int numShards() {
-            return shards.size();
-        }
-
-        @Override
-        public Iterator<ShardRouting> iterator() {
-            return shards.iterator();
-        }
-
-        public void removeShard(ShardRouting shard) {
-            highestPrimary = -1;
-            assert shards.contains(shard) : "Shard not allocated on current node: " + shard;
-            shards.remove(shard);
-        }
-
-        public void addShard(ShardRouting shard) {
-            highestPrimary = -1;
-            assert !shards.contains(shard) : "Shard already allocated on current node: " + shard;
-            shards.add(shard);
-        }
-
-        public boolean containsShard(ShardRouting shard) {
-            return shards.contains(shard);
-        }
     }
 
     static final class NodeSorter extends IntroSorter {
